ID-126 - replaced MarthaResolver with a mocked DrsHubResolver
