+++ core/src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/drs/DrsResolver.scala
@@ -5,7 +5,9 @@ import org.broadinstitute.dsde.rawls.model.UserInfo
 import scala.concurrent.Future
 
 trait DrsResolver {
-  val dosDrsUriPattern: String = "^(dos|drs)://.*"
-
   def drsServiceAccountEmail(drsUrl: String, userInfo: UserInfo): Future[Option[String]]
+}
+
+object DrsResolver {
+  val dosDrsUriPattern: String = "^(dos|drs)://.*"
 }
