+++ core/src/main/scala/org/broadinstitute/dsde/rawls/jobexec/WorkflowSubmissionActor.scala
@@ -362,10 +362,10 @@ trait WorkflowSubmission extends FutureSupport with LazyLogging with MethodWiths
       inputResolutions <- workflow.inputResolutions
       attribute <- inputResolutions.value.toSeq // toSeq makes the for comp work
       dosAttributeValue <- attribute match {
-        case AttributeString(s) if s.matches(drsResolver.dosDrsUriPattern) => Seq(s)
+        case AttributeString(s) if s.matches(DrsResolver.dosDrsUriPattern) => Seq(s)
         case AttributeValueList(valueList) =>
           valueList.collect {
-            case AttributeString(s) if s.value.matches(drsResolver.dosDrsUriPattern) => s
+            case AttributeString(s) if s.value.matches(DrsResolver.dosDrsUriPattern) => s
           }
         case _ => Seq.empty
       }
